make some change to change_state to reflect more accurate error messages
